backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.4.0
mutagen==1.47.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
psycopg2-binary>=2.9.10
cryptography
